Fix the relative file path

+++ src/main/java/com/microsoft/jenkins/acs/commands/MarathonDeploymentCommand.java
@@ -11,7 +11,6 @@ import com.microsoft.jenkins.acs.util.Constants;
 import com.microsoft.jenkins.acs.util.JSchClient;
 import com.microsoft.jenkins.acs.util.JsonHelper;
 import hudson.FilePath;
-import hudson.Launcher;
 
 import java.io.File;
 import java.io.IOException;
@@ -23,26 +22,33 @@ public class MarathonDeploymentCommand implements ICommand<MarathonDeploymentCom
         String host = context.getMgmtFQDN();
         SSHUserPrivateKey sshCredentials = context.getSshCredentials();
         String linuxAdminUsername = context.getLinuxAdminUsername();
-
-        String marathonConfigFile = context.getConfigFilePaths();
+        String relativeFilePath = context.getConfigFilePaths();
 
         JSchClient client = null;
         try {
+            FilePath[] configPaths = context.getWorkspace().list(relativeFilePath);
+            if (configPaths == null || configPaths.length == 0) {
+                context.logError("No configuration found at: " + relativeFilePath);
+                context.setDeploymentState(DeploymentState.UnSuccessful);
+                return;
+            }
+
             client = new JSchClient(host, Constants.DCOS_SSH_PORT, linuxAdminUsername, sshCredentials, context);
 
+            for (FilePath configPath : configPaths) {
+                String deployedFilename = "acsDep" + Calendar.getInstance().getTimeInMillis() + ".json";
+                context.logStatus("Copying marathon file to remote file: " + deployedFilename);
+                client.copyTo(new File(configPath.getRemote()), deployedFilename);
 
-            String deployedFilename = "acsDep" + Calendar.getInstance().getTimeInMillis() + ".json";
-            context.logStatus("Copying marathon file to remote file: " + deployedFilename);
-            client.copyTo(new File(marathonConfigFile), deployedFilename);
-
-            String appId = JsonHelper.getId(marathonConfigFile);
-            //ignore if app does not exist
-            context.logStatus(String.format("Deleting application with appId: '%s' if it exists", appId));
-            client.execRemote("curl -X DELETE http://localhost/marathon/v2/apps/" + appId);
-            context.logStatus(String.format("Deploying file '%s' with appId %s to marathon.", deployedFilename, appId));
-            client.execRemote("curl -i -H 'Content-Type: application/json' -d@" + deployedFilename + " http://localhost/marathon/v2/apps");
+                String appId = JsonHelper.getId(configPath.getRemote());
+                //ignore if app does not exist
+                context.logStatus(String.format("Deleting application with appId: '%s' if it exists", appId));
+                client.execRemote("curl -X DELETE http://localhost/marathon/v2/apps/" + appId);
+                context.logStatus(String.format("Deploying file '%s' with appId %s to marathon.", deployedFilename, appId));
+                client.execRemote("curl -i -H 'Content-Type: application/json' -d@" + deployedFilename + " http://localhost/marathon/v2/apps");
+            }
             context.setDeploymentState(DeploymentState.Success);
-        } catch (JSchException | IOException e) {
+        } catch (JSchException | IOException | InterruptedException e) {
             context.logError("Error deploying application to marathon:", e);
             context.setDeploymentState(DeploymentState.UnSuccessful);
         } finally {
@@ -60,8 +66,6 @@ public class MarathonDeploymentCommand implements ICommand<MarathonDeploymentCom
         SSHUserPrivateKey getSshCredentials();
 
         String getConfigFilePaths();
-
-        Launcher getLauncher();
 
         FilePath getWorkspace();
     }
